fixed DUP error

+++ muCNV/genotyper.cpp
@@ -430,7 +430,7 @@ void Genotyper::call_cnv(sv &S, SvData& D, SvGeno &G)
         }
     }
 
-	double callrate = G.ns / n_sample;
+	double callrate = (double)G.ns / n_sample;
 
     if ((G.dp_flag || G.dp2_flag || G.read_flag ) && callrate>0.5 && G.ac>0)
         G.b_pass = true;
